requirements.txt
pydub~=0.25.1
gTTS~=2.5.1
discord.py
PyNaCl==1.5.0
discord.py==2.3.2
